pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pybackup"
version = "1.0.0"
description = "Production-grade backup engine for files and databases with a web UI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "PyBackup Contributors" }]
keywords = ["backup", "database", "mongodb", "postgres", "mysql", "mssql", "files"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Archiving :: Backup",
]

dependencies = [
    "click>=8.1",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "black", "ruff", "mypy"]

[project.scripts]
pybackup = "pybackup.cli:main"

[project.urls]
Homepage = "https://github.com/selvaraj2003/pybackup"
Documentation = "https://github.com/selvaraj2003/pybackup"
Repository = "https://github.com/selvaraj2003/pybackup"
Issues = "https://github.com/selvaraj2003/pybackup"

[tool.setuptools.packages.find]
where = ["."]
include = ["pybackup*"]

[tool.setuptools.package-data]
"pybackup.static" = ["**/*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
